requirements.txt
abstract_singleton==1.0.1
aenum==3.1.16
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.1
annotated-types==0.7.0
anyio==3.6.2
async-generator==1.10
async-timeout==4.0.2
asynctest==0.13.0
attrs==23.1.0
auto_gpt_plugin_template==0.0.3
autoflake==2.1.1
beautifulsoup4==4.12.2
black==23.3.0
blinker==1.9.0
blis==0.7.9
cachetools==5.3.0
catalogue==2.0.8
certifi==2024.8.30
cffi==1.15.1
cfgv==3.3.1
chardet==4.0.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
confection==0.0.4
coverage==7.2.5
cryptography==40.0.2
cssselect==1.2.0
cymem==2.0.7
Deprecated==1.2.18
distlib==0.3.6
distro==1.8.0
dnspython==2.3.0
docker==6.1.2
duckduckgo-search==3.0.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl#sha256=0964370218b7e1672a30ac50d72cdc6b16f7c867496f1d60925691188f4d2510
et-xmlfile==1.1.0
exceptiongroup==1.1.1
execnet==1.9.0
filelock==3.12.0
flake8==6.0.0
Flask==3.1.1
frozenlist==1.3.3
future==1.0.0
ghp-import==2.1.0
gitdb==4.0.10
GitPython==3.1.31
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.18.1
google-auth-httplib2==0.1.0
googleapis-common-protos==1.59.0
gTTS==2.3.1
h11==0.14.0
httpcore==0.17.1
httplib2==0.22.0
httpx==0.24.1
identify==2.5.24
idna==2.10
importlib-metadata==6.6.0
iniconfig==2.0.0
isort==5.12.0
itsdangerous==2.2.0
Jinja2==3.1.2
jsonschema==4.17.3
keyboard==0.13.5
langcodes==3.3.0
line-bot-sdk==3.17.1
loguru==0.7.0
lxml==4.9.2
Markdown==3.3.7
MarkupSafe==2.1.2
mccabe==0.7.0
mergedeep==1.3.4
mkdocs==1.4.3
MouseInfo==0.1.3
mss==9.0.1
multidict==6.0.4
murmurhash==1.0.9
mypy==1.3.0
mypy-extensions==1.0.0
ninja==1.11.1
nodeenv==1.8.0
numpy==1.24.3
oauthlib==3.2.2
openai==0.27.2
openapi-python-client==0.13.4
opencv-python==4.7.0.72
openpyxl==3.0.7
orjson==3.8.10
outcome==1.2.0
packaging==23.1
pathspec==0.11.1
pathy==0.10.1
Pillow==9.5.0
pinecone-client==2.2.1
platformdirs==3.5.1
playsound==1.2.2
pluggy==1.0.0
pre-commit==3.3.2
preshed==3.0.8
propcache==0.3.2
protobuf==4.23.1
py-cpuinfo==9.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAutoGUI==0.9.53
pycodestyle==2.10.0
pycparser==2.21
pydantic==2.11.7
pydantic_core==2.33.2
pyflakes==3.0.1
pygame==2.4.0
PyGetWindow==0.0.9
pymdown-extensions==10.0.1
pymongo==3.12.0
PyMsgBox==1.0.9
pymunk==6.4.0
pyobjc==9.1.1
pyobjc-core==9.1.1
pyobjc-framework-Accounts==9.1.1
pyobjc-framework-AddressBook==9.1.1
pyobjc-framework-AppleScriptKit==9.1.1
pyobjc-framework-AppleScriptObjC==9.1.1
pyobjc-framework-ApplicationServices==9.1.1
pyobjc-framework-AudioVideoBridging==9.1.1
pyobjc-framework-Automator==9.1.1
pyobjc-framework-AVFoundation==9.1.1
pyobjc-framework-AVKit==9.1.1
pyobjc-framework-CalendarStore==9.1.1
pyobjc-framework-CFNetwork==9.1.1
pyobjc-framework-CloudKit==9.1.1
pyobjc-framework-Cocoa==9.1.1
pyobjc-framework-Collaboration==9.1.1
pyobjc-framework-Contacts==9.1.1
pyobjc-framework-ContactsUI==9.1.1
pyobjc-framework-CoreAudio==9.1.1
pyobjc-framework-CoreAudioKit==9.1.1
pyobjc-framework-CoreBluetooth==9.1.1
pyobjc-framework-CoreData==9.1.1
pyobjc-framework-CoreLocation==9.1.1
pyobjc-framework-CoreMedia==9.1.1
pyobjc-framework-CoreMediaIO==9.1.1
pyobjc-framework-CoreMIDI==9.1.1
pyobjc-framework-CoreServices==9.1.1
pyobjc-framework-CoreText==9.1.1
pyobjc-framework-CoreWLAN==9.1.1
pyobjc-framework-CryptoTokenKit==9.1.1
pyobjc-framework-DictionaryServices==9.1.1
pyobjc-framework-DiscRecording==9.1.1
pyobjc-framework-DiscRecordingUI==9.1.1
pyobjc-framework-DiskArbitration==9.1.1
pyobjc-framework-DVDPlayback==9.1.1
pyobjc-framework-EventKit==9.1.1
pyobjc-framework-ExceptionHandling==9.1.1
pyobjc-framework-FinderSync==9.1.1
pyobjc-framework-FSEvents==9.1.1
pyobjc-framework-GameCenter==9.1.1
pyobjc-framework-GameController==9.1.1
pyobjc-framework-GameKit==9.1.1
pyobjc-framework-GameplayKit==9.1.1
pyobjc-framework-ImageCaptureCore==9.1.1
pyobjc-framework-IMServicePlugIn==9.1.1
pyobjc-framework-InputMethodKit==9.1.1
pyobjc-framework-InstallerPlugins==9.1.1
pyobjc-framework-InstantMessage==9.1.1
pyobjc-framework-Intents==9.1.1
pyobjc-framework-IOBluetooth==9.1.1
pyobjc-framework-IOBluetoothUI==9.1.1
pyobjc-framework-IOSurface==9.1.1
pyobjc-framework-iTunesLibrary==9.1.1
pyobjc-framework-LatentSemanticMapping==9.1.1
pyobjc-framework-LaunchServices==9.1.1
pyobjc-framework-libdispatch==9.1.1
pyobjc-framework-libxpc==9.1.1
pyobjc-framework-LocalAuthentication==9.1.1
pyobjc-framework-MapKit==9.1.1
pyobjc-framework-MediaAccessibility==9.1.1
pyobjc-framework-MediaLibrary==9.1.1
pyobjc-framework-MediaPlayer==9.1.1
pyobjc-framework-MediaToolbox==9.1.1
pyobjc-framework-Metal==9.1.1
pyobjc-framework-MetalKit==9.1.1
pyobjc-framework-ModelIO==9.1.1
pyobjc-framework-MultipeerConnectivity==9.1.1
pyobjc-framework-NetFS==9.1.1
pyobjc-framework-NetworkExtension==9.1.1
pyobjc-framework-NotificationCenter==9.1.1
pyobjc-framework-OpenDirectory==9.1.1
pyobjc-framework-OSAKit==9.1.1
pyobjc-framework-Photos==9.1.1
pyobjc-framework-PhotosUI==9.1.1
pyobjc-framework-PreferencePanes==9.1.1
pyobjc-framework-PubSub==9.1.1
pyobjc-framework-Quartz==9.1.1
pyobjc-framework-SafariServices==9.1.1
pyobjc-framework-SceneKit==9.1.1
pyobjc-framework-ScreenSaver==9.1.1
pyobjc-framework-ScriptingBridge==9.1.1
pyobjc-framework-SearchKit==9.1.1
pyobjc-framework-Security==9.1.1
pyobjc-framework-SecurityFoundation==9.1.1
pyobjc-framework-SecurityInterface==9.1.1
pyobjc-framework-ServiceManagement==9.1.1
pyobjc-framework-Social==9.1.1
pyobjc-framework-SpriteKit==9.1.1
pyobjc-framework-StoreKit==9.1.1
pyobjc-framework-SyncServices==9.1.1
pyobjc-framework-SystemConfiguration==9.1.1
pyobjc-framework-VideoToolbox==9.1.1
pyobjc-framework-WebKit==9.1.1
pyOpenSSL==23.1.1
pyparsing==3.0.9
pyperclip==1.8.2
PyRect==0.2.0
pyrsistent==0.19.3
PyScreeze==0.1.28
PySocks==1.7.1
pytest==7.3.1
pytest-asyncio==0.21.0
pytest-benchmark==4.0.0
pytest-cov==4.0.0
pytest-integration==0.2.3
pytest-mock==3.10.0
pytest-recording==0.12.2
pytest-xdist==3.3.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytweening==1.0.7
PyYAML==6.0
pyyaml_env_tag==0.1
readability-lxml==0.8.1
redis==4.5.5
regex==2023.5.5
requests==2.32.4
requests-oauthlib==1.3.1
rsa==4.9
rubicon-objc==0.4.6
selenium==4.1.4
shellingham==1.5.0.post1
six==1.16.0
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
spacy==3.5.3
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.6
thinc==8.1.10
tiktoken==0.3.3
tomli==2.0.1
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
tweepy==4.14.0
typer==0.7.0
typing-inspection==0.4.1
typing_extensions==4.14.0
uritemplate==4.1.1
urllib3==2.5.0
vcrpy==4.2.1
virtualenv==20.23.0
wasabi==1.1.1
watchdog==3.0.0
webdriver-manager==3.8.6
websocket-client==1.5.1
Werkzeug==3.1.3
wrapt==1.15.0
wsproto==1.2.0
yarl==1.20.1
zipp==3.15.0
